Moving policy components in to the policy editor folder so that it can be moved to shared
