+++ frontend/resourceadm/components/PolicyEditor/ExpandablePolicyCard/ExpandablePolicyCard.module.css
