+++ frontend/resourceadm/components/PolicyEditor/ExpandablePolicyCard/index.ts
