+++ frontend/resourceadm/components/PolicyEditor/ExpandablePolicyElement/DropdownMenu.module.css
