+++ frontend/resourceadm/components/PolicyEditor/ExpandablePolicyElement/ExpandablePolicyElement.module.css
