+++ frontend/resourceadm/components/PolicyEditor/ExpandablePolicyElement/ExpandablePolicyElement.tsx
