+++ frontend/resourceadm/components/PolicyEditor/ResourceNarrowingList/ResourceNarrowingList.module.css
