+++ frontend/resourceadm/components/PolicyEditor/ResourceNarrowingList/ResourceNarrowingList.tsx
