+++ frontend/resourceadm/components/PolicyEditor/ResourceNarrowingList/index.ts
